Preload fonts. Move /routes to /pages.
