+++ package.json
@@ -17,6 +17,9 @@
     "react-scripts": "latest",
     "typescript": "latest"
   },
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
