+++ public/index.html
@@ -1,23 +1,34 @@
 <!doctype html>
-<html lang="en">
+<html lang='en'>
   <head>
-    <meta charset="utf-8" />
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="initial-scale=1, width=device-width" />
+    <meta charset='utf-8' />
+    <link rel='shortcut icon' href='%PUBLIC_URL%/favicon.ico' />
+    <meta name='viewport' content='initial-scale=1, width=device-width' />
     <title>Mike Young's Portfolio</title>
-    <!-- Fonts to support Material Design -->
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap"
-    />
-    <link rel='preconnect' href='https://fonts.googleapis.com' />
-    <link rel='preconnect' href='https://fonts.gstatic.com' crossOrigin='anonymous' />
-    <link href='https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap' rel='stylesheet'></link>
+
+    <!-- Load custom fonts -->
+    <link rel='preload' href='./fonts/Montserrat-Regular.woff2' as='font' type='font/woff2' crossorigin />
+    <link rel='preload' href='./fonts/Montserrat-Italic.woff2' as='font' type='font/woff2' crossorigin />
+    <style>
+      @font-face {
+        font-family: 'Montserrat';
+        src: url('./fonts/Montserrat-Regular.woff2') format('woff2');
+        font-weight: 100 900;
+        font-style: normal;
+        font-display: swap;
+      }
+
+      @font-face {
+        font-family: 'Montserrat';
+        src: url('./fonts/Montserrat-Italic.woff2') format('woff2');
+        font-weight: 100 900;
+        font-style: italic;
+        font-display: swap;
+      }
+    </style>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id='root'></div>
   </body>
 </html>
